After a thousand years.... it compiles again! However, it doesn't run yet. We'll meet again tomorrow.

+++ framework/include/buffer_data.h
@@ -69,7 +69,7 @@ namespace cgb
 		size_t mSizeOfOneElement;
 
 		// The number of indices in the buffer
-		uint32_t mIndexCount;
+		size_t mIndexCount;
 	};
 
 }
